[Update] 적이 죽은 후 5초뒤 gameObject.SetActive(false)가 되도록 수정

+++ Assets/Scripts/Item/Consumable/NormalDrone.cs
@@ -7,7 +7,7 @@ public class NormalDrone : ItemConsumable
     {
         base.UseConsumable();
 
-        GameObject go = ObjectPoolingManager.Instance.GetGameObject((int)ObjectPoolType.Drone);
+        GameObject go = ObjectPoolingManager.Instance.GetGameObject(ObjectPoolType.Drone);
         Drone drone = go.GetComponent<Drone>();
         drone.ActiveDrone(30f);
