+++ Assets/Scripts/Item/Consumable/RareDrone.cs
@@ -5,7 +5,7 @@ public class RareDrone : ItemConsumable
     public override void UseConsumable()
     {
         base.UseConsumable();
-        GameObject go = ObjectPoolingManager.Instance.GetGameObject((int)ObjectPoolType.Drone);
+        GameObject go = ObjectPoolingManager.Instance.GetGameObject(ObjectPoolType.Drone);
         Drone drone = go.GetComponent<Drone>();
         drone.ActiveDrone(60f);
     }
