+++ Assets/Scripts/Item/Consumable/UniqueDrone.cs
@@ -5,7 +5,7 @@ public class UniqueDrone : ItemConsumable
     public override void UseConsumable()
     {
         base.UseConsumable();
-        GameObject go = ObjectPoolingManager.Instance.GetGameObject((int)ObjectPoolType.Drone);
+        GameObject go = ObjectPoolingManager.Instance.GetGameObject(ObjectPoolType.Drone);
         Drone drone = go.GetComponent<Drone>();
         drone.ActiveDrone(100f);
     }
